# wfc/canvas/static/scripts/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "ContractSpec": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "required": {
            "default": true,
            "title": "Required",
            "type": "boolean"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "value_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value Type"
          }
        },
        "required": [
          "type",
          "name"
        ],
        "title": "ContractSpec",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MethodRegisterRequest": {
        "properties": {
          "directory": {
            "title": "Directory",
            "type": "string"
          },
          "method_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Method Name"
          },
          "module": {
            "title": "Module",
            "type": "string"
          }
        },
        "required": [
          "directory",
          "module"
        ],
        "title": "MethodRegisterRequest",
        "type": "object"
      },
      "MethodValidateRequest": {
        "properties": {
          "method": {
            "title": "Method",
            "type": "string"
          },
          "module": {
            "title": "Module",
            "type": "string"
          }
        },
        "required": [
          "module",
          "method"
        ],
        "title": "MethodValidateRequest",
        "type": "object"
      },
      "ModuleRegisterRequest": {
        "properties": {
          "contracts": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ContractSpec"
            },
            "title": "Contracts",
            "type": "array"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "folder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "ModuleRegisterRequest",
        "type": "object"
      },
...
      "SampleRegisterRequest": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "registration_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "copy",
            "title": "Registration Mode"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "name",
          "source"
        ],
        "title": "SampleRegisterRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Workflow Canvas",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/fs/browse": {
      "get": {
        "description": "List directory contents under the project root.\n\n``path`` is relative to the project root. Defaults to the root itself.\nRejects any path that resolves outside the project root. Safe for a\nlocalhost single-user dev tool: the server is already scoped to the\nproject the user chose to launch it against.",
        "operationId": "fs_browse_api_fs_browse_get",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": false,
            "schema": {
              "default": "",
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fs Browse"
      }
    },
...
    "/api/registry/methods": {
      "get": {
        "description": "List registered methods for the Registry tab.\n\nShape: `{methods: [{name, module, env, validated, runCount, source}]}`.\n\n`validated` is a `bool | null` (null = never checked); this list does not\nread the validation cache, so it is always null.",
        "operationId": "get_registry_methods_api_registry_methods_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Registry Methods"
      },
      "post": {
        "description": "Wrap ``wfc.registration.register_method`` with optional dry-run preflight.",
        "operationId": "register_method_endpoint_api_registry_methods_post",
        "parameters": [
          {
            "in": "query",
            "name": "dryRun",
            "required": false,
            "schema": {
              "default": false,
              "title": "Dryrun",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MethodRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Method Endpoint"
      }
    },
    "/api/registry/methods/validate": {
      "post": {
        "description": "Re-validate an existing method's env+import. Caches by fingerprint.",
        "operationId": "validate_registry_method_api_registry_methods_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MethodValidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Registry Method"
      }
    },
    "/api/registry/methods/{module_name}/{method_name}/detail": {
      "get": {
        "description": "Read-only view of a method's directory + its parsed contract.\n\nLists every file in the method dir (no recursion) with its content and a\nlanguage hint for syntax highlighting. Returns the DB-parsed contract\n(input_slots, output_slots, params_schema) alongside so the frontend can\nrender it as structured tables without re-parsing YAML.",
        "operationId": "get_registry_method_detail_api_registry_methods__module_name___method_name__detail_get",
        "parameters": [
          {
            "in": "path",
            "name": "module_name",
            "required": true,
            "schema": {
              "title": "Module Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "method_name",
            "required": true,
            "schema": {
              "title": "Method Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Registry Method Detail"
      }
    },
    "/api/registry/modules": {
      "get": {
        "description": "List registered modules for the Registry tab.\n\nShape: `{modules: [{name, description, contracts[], methods, source}]}`.\n`color` is intentionally omitted -- the frontend assigns it from the\n`MOD_COLORS` cycle by registration order.",
        "operationId": "get_registry_modules_api_registry_modules_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Registry Modules"
      },
      "post": {
        "description": "Wrap ``wfc.registration.register_module`` with optional dry-run preflight.",
        "operationId": "register_module_endpoint_api_registry_modules_post",
        "parameters": [
          {
            "in": "query",
            "name": "dryRun",
            "required": false,
            "schema": {
              "default": false,
              "title": "Dryrun",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModuleRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Module Endpoint"
      }
    },
    "/api/registry/samples": {
      "get": {
        "description": "List registered samples for the Registry tab.\n\nShape: `{samples: [{name, source, size, hash, pushed, runCount,\nregistered_at}]}`.\n\n`pushed` is `True` when a DVC content hash is present (indicates the sample\nis in the DVC cache and pushable to the remote); `False` otherwise.",
        "operationId": "get_registry_samples_api_registry_samples_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Registry Samples"
      },
      "post": {
        "description": "Wrap ``wfc.registration.register_sample`` behind an HTTP endpoint.",
        "operationId": "register_sample_endpoint_api_registry_samples_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SampleRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Sample Endpoint"
      }
    },
...
  }
}